+++ R/lotrrs.R
@@ -3,8 +3,9 @@
 #' Estimates a ratio of relative risk surfaces and computes the asymptotic p-value surface for a single gate with two conditions. Includes features for basic visualization. This function is used internally within the \code{\link{gating}} function to extract the points within the significant areas. This function can also be used as a standalone function.
 #'
 #' @param dat Input data frame flow cytometry data with five (5) features (columns): 1) ID, 2) Condition A ID, 3) Condition B ID, 4) Marker A as x-coordinate, 5) Marker B as y-coordinate.
+#' @param bandw Optional, numeric. Fixed bandwidth for the kernel density estimation. Default is based on the internal \code{[sparr]{OS}} function.
 #' @param alpha Numeric. The two-tailed alpha level for significance threshold (default is 0.05).
-#' @param p_correct Optional. Character string specifying whether to apply a correction for multiple comparisons including a False Discovery Rate \code{p_correct = "FDR"}, a spatially dependent Sidak correction \code{p_correct = "correlated Sidak"}, a spatially dependent Bonferroni correction \code{p_correct = "correlated Bonferroni"}, an independent Sidak correction \code{p_correct = "uncorrelated Sidak"}, and an independent Bonferroni correction \code{p_correct = "uncorrelated Bonferroni"}. If \code{p_correct = "none"} (the default), then no correction is applied. 
+#' @param p_correct Optional. Character string specifying whether to apply a correction for multiple comparisons including a False Discovery Rate \code{p_correct = "FDR"}, a spatially dependent Sidak correction \code{p_correct = "correlated Sidak"}, a spatially dependent Bonferroni correction \code{p_correct = "correlated Bonferroni"}, an independent Sidak correction \code{p_correct = "uncorrelated Sidak"}, an independent Bonferroni correction \code{p_correct = "uncorrelated Bonferroni"}, and a correction based on Random Field Theory using an equation by Adler and Hasofer \code{p_correct = "Adler and Hasofer"} or an equation by Friston et al. \code{p_correct = "Friston"}. If \code{p_correct = "none"} (the default), then no correction is applied. 
 #' @param nbc Optional. An integer for the number of bins when \code{p_correct = "correlated"}. Similar to \code{nbclass} argument in \code{\link[SpatialPack]{modified.ttest}}. The default is 30. 
 #' @param plot_gate Logical. If \code{TRUE}, the output includes basic data visualization.
 #' @param save_gate Logical. If \code{TRUE}, the output saves the visualization as a separate PNG file.
@@ -16,7 +17,7 @@
 #' @param c1n Optional, character. The name of the level for the numerator of condition A. The default is null and the first level is treated as the numerator. 
 #' @param c2n Optional, character. The name of the level for the numerator of condition B. The default is null and the first level is treated as the numerator.
 #' @param win Optional. Object of class \code{owin} for a custom two-dimensional window within which to estimate the surfaces. The default is NULL and calculates a convex hull around the data. 
-#' @param ... Arguments passed to \code{\link[sparr]{risk}} to select bandwidth, edge correction, and resolution.
+#' @param ... Arguments passed to \code{\link[sparr]{risk}} to select resolution.
 #'
 #' @details This function estimates a ratio of relative risk surfaces and computes the asymptotic p-value surface for a single gate with two conditions using three successive \code{\link[sparr]{risk}} functions. A relative risk surface is estimated for Condition A at each level of Condition B and then a ratio of the two relative risk surfaces is computed. 
 #' 
@@ -26,7 +27,7 @@
 #' 
 #' The p-value surface of the ratio of relative risk surfaces is estimated assuming asymptotic normality of the ratio value at each gridded knot. The bandwidth is fixed across all layers. Basic visualization is available if \code{plot_gate = TRUE}. 
 #' 
-#' Provides functionality for a correction for multiple testing. If \code{p_correct = "FDR"}, calculates a False Discovery Rate by Benjamini and Hochberg. If \code{p_correct = "uncorrelated Sidak"}, calculates an independent Sidak correction. If \code{p_correct = "uncorrelated Bonferroni"}, calculates an independent Bonferroni correction. If \code{p_correct = "correlated Sidak"} or if \code{p_correct = "correlated Bonferroni"}, then the corrections take into account the into account the spatial correlation of the surface. (NOTE: If \code{p_correct = "correlated Sidak"} or if \code{p_correct = "correlated Bonferroni"}, it may take a considerable amount of computation resources and time to calculate). If \code{p_correct = "none"} (the default), then the function does not account for multiple testing and uses the uncorrected \code{alpha} level. See the internal \code{pval_correct} function documentation for more details.
+#' Provides functionality for a correction for multiple testing. If \code{p_correct = "FDR"}, calculates a False Discovery Rate by Benjamini and Hochberg. If \code{p_correct = "uncorrelated Sidak"}, calculates an independent Sidak correction. If \code{p_correct = "uncorrelated Bonferroni"}, calculates an independent Bonferroni correction. If \code{p_correct = "correlated Sidak"} or if \code{p_correct = "correlated Bonferroni"}, then the corrections take into account the into account the spatial correlation of the surface. (NOTE: If \code{p_correct = "correlated Sidak"} or if \code{p_correct = "correlated Bonferroni"}, it may take a considerable amount of computation resources and time to calculate). If \code{p_correct = "Adler and Hasofer"} or if \code{p_correct = "Friston"}, then calculates a correction based on Random Field Theory. If \code{p_correct = "none"} (the default), then the function does not account for multiple testing and uses the uncorrected \code{alpha} level. See the internal \code{pval_correct} function documentation for more details.
 #' 
 #' The two condition variables (Condition A and Condition B) within \code{dat} must be of class 'factor' with two levels. The first level in each variable is considered the numerator (i.e., "case") value and the second level is considered the denominator (i.e., "control") value. The levels can also be specified using the \code{c1n} and \code{c2n} parameters.
 #'
@@ -54,6 +55,7 @@
 #' test_lotrrs <- lotrrs(dat = randCyto)
 #' 
 lotrrs <- function(dat, 
+                   bandw = NULL,
                    alpha = 0.05, 
                    p_correct = "none",
                    nbc = NULL,
@@ -79,7 +81,7 @@ lotrrs <- function(dat,
   if (nlevels(dat[ , 3]) != 2) { stop("The third feature of 'dat' must be a binary factor") }
   
   ## p_correct
-    match.arg(p_correct, choices = c("none", "FDR", "correlated Sidak", "correlated Bonferroni", "uncorrelated Sidak", "uncorrelated Bonferroni"))
+    match.arg(p_correct, choices = c("none", "FDR", "correlated Sidak", "correlated Bonferroni", "uncorrelated Sidak", "uncorrelated Bonferroni", "Adler and Hasofer", "Friston"))
   
   ## alpha
   if (alpha <= 0 | alpha >= 1 ) {
@@ -140,16 +142,17 @@ lotrrs <- function(dat,
                                                                     window = win)))
   
   # Estimate two SRRs
-  both_h0 <- sparr::OS(both_ppp, "geometric")
+  if (is.null(bandw)){ bandw <- sparr::OS(both_ppp, "geometric") }
+  
   rm(both_ppp) # conserve memory
   denom_rr <- sparr::risk(denom_ppp,
-                          h0 = both_h0,
+                          h0 = bandw,
                           log = FALSE,
                           edge = "diggle",
                           verbose = FALSE,
                           ...)
   numer_rr <- sparr::risk(numer_ppp,
-                          h0 = both_h0,
+                          h0 = bandw,
                           log = FALSE,
                           edge = "diggle",
                           verbose = FALSE,
@@ -157,12 +160,12 @@ lotrrs <- function(dat,
   
   # Create objects of class 'bivden'
   denom_bd <- sparr::bivariate.density(denom_ppp,
-                                       h0 = both_h0,
+                                       h0 = bandw,
                                        edge = "diggle",
                                        verbose = FALSE,
                                        ...)
   numer_bd <- sparr::bivariate.density(numer_ppp,
-                                       h0 = both_h0,
+                                       h0 = bandw,
                                        edge = "diggle",
                                        verbose = FALSE,
                                        ...)
@@ -192,7 +195,7 @@ lotrrs <- function(dat,
   ## Also estimate the asymptotic p-value surface
   suppressMessages(suppressWarnings(out <- sparr::risk(f = numer_rr_bd,
                                                        g = denom_rr_bd,
-                                                       h0 = both_h0,
+                                                       h0 = bandw,
                                                        tolerate = TRUE,
                                                        edge = "diggle",
                                                        verbose = FALSE,
@@ -209,27 +212,32 @@ lotrrs <- function(dat,
   }
   
   # Alpha level
-  if (p_correct == "none") { alpha_correct <- alpha }
+  if (p_correct == "none") { p_critical <- alpha }
   if (p_correct == "FDR") { 
-    alpha_correct <- pval_correct(input = out, type = "FDR", alpha = alpha, nbc = nbc)
+    p_critical <- pval_correct(input = out, type = "FDR", alpha = alpha, nbc = nbc)
   }
   if (p_correct == "correlated Sidak") {
     message("Please be patient... Calculating spatially dependent Sidak correction")
-    alpha_correct <- pval_correct(input = out, type = "correlated Sidak", alpha = alpha, nbc = nbc)
+    p_critical <- pval_correct(input = out, type = "correlated Sidak", alpha = alpha, nbc = nbc)
   } 
   if (p_correct == "correlated Bonferroni") {
     message("Please be patient... Calculating spatially dependent Bonferroni correction")
-    alpha_correct <- pval_correct(input = out, type = "correlated Bonferroni", alpha = alpha, nbc = nbc)
+    p_critical <- pval_correct(input = out, type = "correlated Bonferroni", alpha = alpha, nbc = nbc)
   }
   if (p_correct == "uncorrelated Sidak") { 
-    alpha_correct <- pval_correct(input = out, type = "uncorrelated Sidak", alpha = alpha, nbc = nbc)
+    p_critical <- pval_correct(input = out, type = "uncorrelated Sidak", alpha = alpha, nbc = nbc)
   }
-  
   if (p_correct == "uncorrelated Bonferroni") { 
-    alpha_correct <- pval_correct(input = out, type = "uncorrelated Bonferroni", alpha = alpha, nbc = nbc)
+    p_critical <- pval_correct(input = out, type = "uncorrelated Bonferroni", alpha = alpha, nbc = nbc)
+  }
+  if (p_correct == "Adler and Hasofer") { 
+    p_critical <- pval_correct(input = out, type = "Adler and Hasofer", alpha = alpha, nbc = nbc)
+  }
+  if (p_correct == "Friston") { 
+    p_critical <- pval_correct(input = out, type = "Friston", alpha = alpha, nbc = nbc)
   }
 
-  out$alpha <- alpha_correct
+  out$alpha <- p_critical
   
   if (plot_gate == TRUE) {
     # Graphics
@@ -288,7 +296,7 @@ lotrrs <- function(dat,
                                         labels = tr_plot$labels,
                                         cex.axis = 0.67),
                        main = paste("log of the\nrelative risk surfaces\n(bandwidth:",
-                                    round(both_h0, digits = 3),
+                                    round(bandw, digits = 3),
                                     "units)",
                                     sep = " "))
     par(bg = "transparent")