+++ NEWS.md
@@ -2,8 +2,9 @@
 
 # gateR v0.1.7
   * Updated {spatstat} dependency to new packages based on feedback from the Spatstat Team (Adrian Baddeley and Ege Rubak). {spatstat.geom} replaces {spatstat} in Imports
-  * Added additional multiple testing corrections, including False Discovery Rate, spatially dependent Sidak correction, and independent Sidak correction
-  * Updated the calculation of the spatial correlogram in internal pval_correct() function from the correlog() function in the {pgrimess} package to the modified.ttest() function in the {SpatialPack} package
+  * Added additional multiple testing corrections, including False Discovery Rate, spatially dependent Sidak correction, independent Sidak correction, and two corrections based on Random Field Theory (Adler and Hasofer or Friston).
+  * The latter two corrections required a new argument 'bandw' to be added to gating(), lotrrs(), and rrs() functions to allow users to specify bandwidth for the kernel density estimation.
+  * Updated the calculation of the spatial correlogram in internal pval_correct() function from the correlog() function in the {pgrimess} package to the modified.ttest() function in the {SpatialPack} package.
 
 # gateR v0.1.6
   * Updated URLs in gateR-package.Rd